Split WebAssembly global namespace helpers
